Break out parsing of GitCommandHelpers.GetAllChangedFilesCmd output into GetAllChangedFilesOutputParser
