+++ UnitTests/GitCommands.Tests/Git/GetAllChangedFilesOutputParserTest.TestGetStatusChangedFilesFromString.ForScenario.fatal_error.approved.json
@@ -1,0 +1,65 @@
+﻿[
+    {
+        "Name": "unknown info",
+        "OldName": null,
+        "ErrorMessage": null,
+        "TreeGuid": null,
+        "RenameCopyPercentage": null,
+        "Staged": 2,
+        "IsTracked": false,
+        "IsDeleted": false,
+        "IsChanged": false,
+        "IsNew": true,
+        "IsIgnored": false,
+        "IsRenamed": false,
+        "IsCopied": false,
+        "IsConflict": false,
+        "IsAssumeUnchanged": false,
+        "IsSkipWorktree": false,
+        "IsSubmodule": false,
+        "IsDirty": false,
+        "IsStatusOnly": false
+    },
+    {
+        "Name": "adfs.h",
+        "OldName": null,
+        "ErrorMessage": null,
+        "TreeGuid": null,
+        "RenameCopyPercentage": null,
+        "Staged": 3,
+        "IsTracked": true,
+        "IsDeleted": false,
+        "IsChanged": true,
+        "IsNew": false,
+        "IsIgnored": false,
+        "IsRenamed": false,
+        "IsCopied": false,
+        "IsConflict": false,
+        "IsAssumeUnchanged": false,
+        "IsSkipWorktree": false,
+        "IsSubmodule": false,
+        "IsDirty": false,
+        "IsStatusOnly": false
+    },
+    {
+        "Name": "adfs.h",
+        "OldName": null,
+        "ErrorMessage": null,
+        "TreeGuid": null,
+        "RenameCopyPercentage": null,
+        "Staged": 2,
+        "IsTracked": true,
+        "IsDeleted": false,
+        "IsChanged": true,
+        "IsNew": false,
+        "IsIgnored": false,
+        "IsRenamed": false,
+        "IsCopied": false,
+        "IsConflict": false,
+        "IsAssumeUnchanged": false,
+        "IsSkipWorktree": false,
+        "IsSubmodule": false,
+        "IsDirty": false,
+        "IsStatusOnly": false
+    }
+]
