+++ UnitTests/GitCommands.Tests/Git/GetAllChangedFilesOutputParserTest.TestGetStatusChangedFilesFromString.ForScenario.status_ignored_files.approved.json
