+++ UnitTests/GitCommands.Tests/Git/GetAllChangedFilesOutputParserTest.TestGetStatusChangedFilesFromString.ForScenario.status_staged_files.approved.json
