+++ UnitTests/GitCommands.Tests/Git/GetAllChangedFilesOutputParserTest.TestGetStatusChangedFilesFromString.ForScenario.status_untracked_files.approved.json
