+++ UnitTests/GitCommands.Tests/Submodules/SubmoduleStatusProviderTests.cs
@@ -10,7 +10,6 @@ using GitCommands.Git;
 using GitCommands.Submodules;
 using GitUIPluginInterfaces;
 using NUnit.Framework;
-using NUnit.Framework.Constraints;
 
 namespace GitCommandsTests.Submodules
 {
@@ -768,7 +767,7 @@ namespace GitCommandsTests.Submodules
         {
             var cmd = GitCommandHelpers.GetAllChangedFilesCmd(true, UntrackedFilesMode.Default, noLocks: true);
             var output = module.GitExecutable.GetOutput(cmd);
-            return GitCommandHelpers.GetStatusChangedFilesFromString(module, output);
+            return new GetAllChangedFilesOutputParser(() => module).Parse(output);
         }
     }
 }
